requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.6
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2==2.9.10
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
